Revise model template props + rename to model seed props
