+++ src/controllers/model-seed-props/playtext.js
@@ -1,0 +1,5 @@
+export default {
+	characters: [
+		{}
+	]
+};
