+++ src/controllers/model-seed-props/production.js
@@ -1,0 +1,9 @@
+export default {
+	cast: [
+		{
+			roles: [
+				{}
+			]
+		}
+	]
+};
