+++ src/controllers/model-seed-props/theatre.js
@@ -1,0 +1,5 @@
+export default {
+	subTheatres: [
+		{}
+	]
+};
