--- src/controllers/model-template-props/playtext.js
@@ -1,12 +1,0 @@
-export default {
-	name: '',
-	characters: [
-		{
-			name: '',
-			underlyingName: '',
-			differentiator: '',
-			qualifier: '',
-			group: ''
-		}
-	]
-};
