--- src/controllers/model-template-props/production.js
@@ -1,23 +1,0 @@
-export default {
-	name: '',
-	theatre: {
-		name: '',
-		differentiator: ''
-	},
-	playtext: {
-		name: '',
-		differentiator: ''
-	},
-	cast: [
-		{
-			roles: [
-				{
-					name: '',
-					characterName: '',
-					characterDifferentiator: '',
-					qualifier: ''
-				}
-			]
-		}
-	]
-};
