--- src/controllers/model-template-props/theatre.js
@@ -1,10 +1,0 @@
-export default {
-	name: '',
-	differentiator: '',
-	subTheatres: [
-		{
-			name: '',
-			differentiator: ''
-		}
-	]
-};
